Extracted out a contacts class. Hopefully did this correctly
