+++ test/java/seedu/addressbook/commands/AddCommandTest.java
@@ -103,7 +103,7 @@ public class AddCommandTest {
 
         // TODO: add comparison of tags to person.equals and equality methods to
         // individual fields that compare privacy to simplify this
-        assertEquals(Name.EXAMPLE, p.getName().fullName);
+        assertEquals(Name.EXAMPLE, p.getName().value);
         assertEquals(Phone.EXAMPLE, p.getPhone().value);
         assertTrue(p.getPhone().isPrivate());
         assertEquals(Email.EXAMPLE, p.getEmail().value);
